Ensure 'current' array operations use float to prevent casting error

+++ bluepyefe/ecode/HyperDePol.py
@@ -140,9 +140,9 @@ class HyperDePol(Recording):
         toff = int(self.toff / self.dt)
 
         time = numpy.arange(0.0, self.tend, self.dt)
-        current = numpy.full(time.shape, self.hypamp)
-        current[ton:tmid] += self.amp
-        current[tmid:toff] += self.amp2
+        current = numpy.full(time.shape, numpy.float64(self.hypamp))
+        current[ton:tmid] += numpy.float64(self.amp)
+        current[tmid:toff] += numpy.float64(self.amp2)
 
         return time, current
