+++ bluepyefe/ecode/posCheops.py
@@ -140,7 +140,7 @@ class PosCheops(Recording):
         toff = int(self.toff / self.dt)
 
         time = numpy.arange(0.0, self.tend, self.dt)
-        current = numpy.full(time.shape, self.hypamp)
+        current = numpy.full(time.shape, numpy.float64(self.hypamp))
 
         # First peak
         mid = int(0.5 * (self.ton + self.t1) / self.dt)
